WebVTT file processing merged into base class
